Improve database saving confirmation and Studiengang addition
Added log outputs for each database save action to ensure and track the proper saving of data. This will help in debugging and monitoring the application's function. In addition, introduced a line to add new Studiengang instances to the 'Studiengänge' list in MainUI to maintain an updated list of all Studiengang instances.