Add method to match Prüfungsform from string input
Added a public static method getType in Prüfungsform.java that leverages the existing enum to return a Prüfungsform type based on a string input. This can handle the case when the form text doesn't exactly match the enum name, improving the robustness of our code. The Database.java was also updated to use this new method instead of valueOf method which could throw IllegalArgumentException when improperly used.