Commit message:
Replace Date with LocalDate in PrüfungsUI.java, Prüfungsversuch.java, PraktikumUI.java, Praktikumstermin.java, Utility.java

To improve handling of dates, this commit replaces usage of Date with LocalDate in several files of the Studiumsorganisation application. The switch allows for a more accurate representation of dates, better support for time zones and simplifies the usage of date pickers in UI. Furthermore, the commit introduces null pointer exception handling for the "Datum" column in the UI tables.